change coordinate date type to dbl and update seed values

+++ TrailsApi/Models/TrailMarker.cs
@@ -8,9 +8,9 @@ namespace TrailsApi.Models
 
     public string Name { get; set; }
 
-    public int Longitude { get; set; }
+    public double Longitude { get; set; }
 
-    public int Latitude { get; set; }
+    public double Latitude { get; set; }
 
     public bool isTrailhead { get; set; }
